requirements.txt
kafka-python
transformers
torch
cloudevents
requests